# save code and entrypoint into struct