--- a/Cargo.toml
+++ b/Cargo.toml
@@ -2,7 +2,7 @@
 name = "ssr_rs"
 version = "0.2.3"
 authors = ["Valerio <valerioageno@yahoo.it>"]
-edition = "2018"
+edition = "2021"
 description = "Server side rendering with the v8 engine for parse and evaluate the javascript code"
 readme = "./README.md"
 homepage = "https://github.com/Valerioageno/ssr-rs"