--- a/Cargo.toml
+++ b/Cargo.toml
@@ -46,6 +46,7 @@ rocket = "0.5.0-rc.2"
 
 # Benchmark dependencies
 criterion = "0.3.4"
+env_logger = "0.9.0"
 
 [[bench]]
 name = "benchmark"