--- a/benches/README.md
+++ b/benches/README.md
@@ -1,0 +1,3 @@
+Custom HTTP benchmarks are performed using https://github.com/wg/wrk
+
+?? Give a try to https://github.com/tsliwowicz/go-wrk