--- a/examples/run.rs
+++ b/examples/run.rs
@@ -5,5 +5,5 @@ use std::fs::read_to_string;
 fn main() {
     let source = read_to_string("./client/dist/ssr/index.js").unwrap();
 
-    println!("{}", Ssr::render_to_string(&source, "SSR", None))
+    println!("{}", Ssr::one_shot_render(source, "SSR", None))
 }