This is a first step towards a performance improvement. Following a benchmark comparing the new shared struct data with the old implementation (same logic different composition).

Then new implementation reports a slightly improvement.

## New shared struct (examples/shared_ssr.rs)
```
Running 30s test @ http://127.0.0.1:8080/
  12 threads and 400 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   412.62ms   39.40ms 706.59ms   86.83%
    Req/Sec    79.52     24.95   180.00     69.97%
  28629 requests in 30.09s, 18.32MB read
  Socket errors: connect 0, read 377, write 0, timeout 0
Requests/sec:    951.46
Transfer/sec:    623.47KB
```

## Old implementation like (examples/actix.rs)
```
Running 30s test @ http://127.0.0.1:8080/
  12 threads and 400 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   433.91ms   43.03ms 713.45ms   89.73%
    Req/Sec    75.58     24.76   161.00     64.47%
  27219 requests in 30.09s, 17.42MB read
  Socket errors: connect 0, read 356, write 3, timeout 0
Requests/sec:    904.67
Transfer/sec:    592.80KB
```